+++ lib/sec_id/cfi.rb
@@ -179,6 +179,8 @@ module SecId
     # @return [String, nil] attribute 4 (position 6)
     attr_reader :attr4
 
+    has_check_digit false
+
     # @param cfi [String] the CFI string to parse
     def initialize(cfi)
       cfi_parts = parse(cfi)
@@ -190,11 +192,6 @@ module SecId
       @attr3 = cfi_parts[:attr3]
       @attr4 = cfi_parts[:attr4]
       @check_digit = nil
-    end
-
-    # @return [Boolean] always false - CFI has no check digit
-    def has_check_digit?
-      false
     end
 
     # Validates format including category and group codes.