+++ lib/sec_id/cik.rb
@@ -18,6 +18,8 @@ module SecId
   class CIK < Base
     include Normalizable
 
+    has_check_digit false
+
     # Regular expression for parsing CIK components.
     ID_REGEX = /\A
       (?=\d{1,10}\z)(?<padding>0*)(?<identifier>[1-9]\d{0,9})
@@ -32,11 +34,6 @@ module SecId
       @padding = cik_parts[:padding]
       @identifier = cik_parts[:identifier]
       @check_digit = nil
-    end
-
-    # @return [Boolean] always false
-    def has_check_digit?
-      false
     end
 
     # Normalizes the CIK to a 10-digit zero-padded format.